requirements.txt
fastapi
uvicorn
langchain
sentence-transformers
transformers
torch
faiss-cpu
spacy
datasets
accelerate
peft
bitsandbytes
openai
groq
mlflow
dvc
label-studio
numpy
pandas
scikit-learn


